% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCoxBaselines.R
\name{Baselinehazard_IMP}
\alias{Baselinehazard_IMP}
\title{Baselinehazard_IMP}
\usage{
Baselinehazard_IMP(datWIDE, CovImp, GImp, YRImp, deltaRImp, beta, alpha,
  TransCov, ASSUME)
}
\arguments{
\item{datWIDE}{A data frame with the following columns: 
\itemize{
\item Y_R, the recurrence event/censoring time
\item delta_R, the recurrence event/censoring indicator
\item Y_D, the death event/censoring time 
\item delta_D, the death event/censoring indicator
\item G, the cure status variable. This takes value 1 for known non-cured, 0 for "known" cured and NA for unknown cure status
}}

\item{CovImp}{A list with IMPNUM elements containing the imputations of Cov output from MultiCure}

\item{GImp}{A matrix with IMPNUM elements containing the imputations of G output from MultiCure}

\item{YRImp}{A matrix with IMPNUM elements containing the imputations of Y_R output from MultiCure}

\item{deltaRImp}{A matrix with IMPNUM elements containing the imputations of delta_R output from MultiCure}

\item{beta}{Current estimate of beta}

\item{alpha}{Current estimate of alpha}

\item{TransCov}{a list with elements: Trans13, Trans24, Trans14, Trans34, PNonCure. Each list element is a vector containing the names of the variables in Cov to be used in the model for the corresponding transition. 13 is NonCured -> Recurrence, 24 is Cured -> Death, 14 is NonCured -> Death, 34 is Recurrence -> Death. PNonCure contains the names of the covariates for the logistic regression for P(NonCure).}

\item{ASSUME}{This variables indicates what equality assumptions we are making regarding the 24 and 14 transitions. The possible options are:
\itemize{
\item 'SameHazard': Lambda_14(t) = Lambda_24(t)
\item 'AllSeparate': No restrictions on Lambda_14(t) and Lambda_24(t)
\item  'ProportionalHazard': Lambda_14(t) = Lambda_24(t) exp(Beta0)
\item  'SameBaseHaz': Lambda^0_14(t) = Lambda^0_24(t), No restrictions on beta_14 and beta_24
}}
}
\value{
EST a list containing the estimates for the baseline hazard function (and cumulative baseline hazard function) for the 1->3, 2->4, 1->4, and 3->4 transitions.
}
\description{
The function Baselinehazard_IMP is used to estimate the baseline hazard functions within the MCEM algorithm under COX baseline hazards
}
\author{
Lauren J Beesley, \email{lbeesley@umich.edu}
}
