% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCoxBaselines.R
\name{BaselineHazard_NOIMP}
\alias{BaselineHazard_NOIMP}
\title{BaselineHazard_NOIMP}
\usage{
BaselineHazard_NOIMP(datWIDE, Cov, beta, alpha, TransCov, ASSUME, p)
}
\arguments{
\item{datWIDE}{A data frame with the following columns: 
\itemize{
\item Y_R, the recurrence event/censoring time
\item delta_R, the recurrence event/censoring indicator
\item Y_D, the death event/censoring time 
\item delta_D, the death event/censoring indicator
\item G, the cure status variable. This takes value 1 for known non-cured, 0 for "known" cured and NA for unknown cure status
}}

\item{Cov}{matrix of covariates used in MultiCure (may have missingness)}

\item{beta}{Current estimate of beta}

\item{alpha}{Current estimate of alpha}

\item{TransCov}{a list with elements: Trans13, Trans24, Trans14, Trans34, PNonCure. Each list element is a vector containing the names of the variables in Cov to be used in the model for the corresponding transition. 13 is NonCured -> Recurrence, 24 is Cured -> Death, 14 is NonCured -> Death, 34 is Recurrence -> Death. PNonCure contains the names of the covariates for the logistic regression for P(NonCure).}

\item{ASSUME}{This variables indicates what equality assumptions we are making regarding the 24 and 14 transitions. The possible options are:
\itemize{
\item 'SameHazard': Lambda_14(t) = Lambda_24(t)
\item 'AllSeparate': No restrictions on Lambda_14(t) and Lambda_24(t)
\item  'ProportionalHazard': Lambda_14(t) = Lambda_24(t) exp(Beta0)
\item  'SameBaseHaz': Lambda^0_14(t) = Lambda^0_24(t), No restrictions on beta_14 and beta_24
}}

\item{p}{The current estimate of the E-step weights}
}
\value{
EST a list containing a step function estimate for the CUMULATIVE baseline hazard function for each transition in the following order: 1->3, 2->4, 1->4, 3->4
}
\description{
The function BaselineHazard_NOIMP is used to estimate the baseline hazard functions within the EM algorithm under COX baseline hazards
}
\author{
Lauren J Beesley, \email{lbeesley@umich.edu}
}
