% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCure.R
\name{construct.C}
\alias{construct.C}
\title{Construct adjacent-difference contrast matrix}
\usage{
construct.C(k)
}
\arguments{
\item{k}{Integer number of groups.}
}
\value{
A \\((k-1)\\times k\\) numeric matrix.
}
\description{
Builds a \\((k-1)\\times k\\) matrix with rows \\((1,-1,0,\\dots)\\), \\((0,1,-1,0,\\dots)\\), etc.
}
\keyword{internal}
