% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVIMPUTEINITIALIZE_Example.R
\name{COVIMPUTEINITIALIZE_Example}
\alias{COVIMPUTEINITIALIZE_Example}
\title{COVIMPUTEINITIALIZE_Example}
\usage{
COVIMPUTEINITIALIZE_Example(Cov, CovMissing)
}
\arguments{
\item{Cov}{matrix of covariates, same as in MultiCure}

\item{CovMissing:}{A matrix indicating which elements of Cov are missing.}
}
\value{
Cov a initialized version of the covariate matrix
}
\description{
The function COVIMPUTEINITIALIZE_Example will initialize the missing covariate values. An example function is included in the package, but the user must specify their own function when applying MultiCure to datasets with covariate missingness. This function must have input/output as described below.
}
\details{
The example code included in the package initializes missing covariate X2 in the Multistate cure model example
}
