% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateMultiCure.R
\name{SimulateMultiCure}
\alias{SimulateMultiCure}
\title{SimulateMultiCure}
\usage{
SimulateMultiCure(type = "NoMissingness")
}
\arguments{
\item{type}{Indicates whether we should simulate data with missingness only in cure status, missingness in cure status and covariates, or missingness in cure status and unequal censoring. Possible values include "No Missingness", "CovariateMissingness" and "UnequalCensoring".}
}
\value{
OUT This is a data frame containing
\itemize{
\item Y_R: Event/Censoring Time for Recurrence
\item delta_R: Event/Censoring Indicator for Recurrence
\item Y_D: Event/Censoring Time for Death
\item delta_D: Event/Censoring Indicator for Death
\item G: Non-Cure Status
\item X1: First Covariate
\item X2: Second Covariate
}
}
\description{
This function will create a simulated dataset under a multistate cure model.
}
\examples{
SIMS = SimulateMultiCure(type = "NoMissingness")
}
