% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutcomeImputationCOXNESTEDWEIBULL.R
\name{UNEQUALCENSIMPUTECOXNESTEDWEIB}
\alias{UNEQUALCENSIMPUTECOXNESTEDWEIB}
\title{UNEQUALCENSIMPUTECOXNESTEDWEIB}
\usage{
UNEQUALCENSIMPUTECOXNESTEDWEIB(datWIDE, beta, alpha, ImputeDat, TransCov)
}
\arguments{
\item{datWIDE}{defined as in MultiCure}

\item{beta}{A vector containing the most recent estimates of beta}

\item{alpha}{A vector containing the most recent estimates of alpha}

\item{ImputeDat}{This is a list with the following elements:
\itemize{
\item UnequalCens: A vector taking value 1 if the subject has unequal follow-up. Note: If subject is assumed cured in datWIDE, they are listed as UnequalCens = 0.
\item CovMissing: A matrix indicating which elements of Cov are missing. Not needed for this imputation.
\item CovImp: A list containing a single imputation of Cov
\item GImp: A vector with a recent single imputation of G
\item YRImp: A vector with a recent single imputation of Y_R
\item deltaRImp: A vector with a recent single imputation of delta_R
\item y: The integral of the target kernel over Yr0 to Yd
\item Basehaz13: A matrix containing the estimate of the baseline hazard function for the 1->3 transition specified intervals (not used)
\item Basehaz24: A matrix containing the estimate of the baseline hazard function for the 2->4 transition specified intervals (not used)
\item Basehaz14: A matrix containing the estimate of the baseline hazard function for the 1->4 transition specified intervals (not used)
\item Basehaz34: A matrix containing the estimate of the baseline hazard function for the 3->4 transition specified intervals (not used)
}}

\item{TransCov}{defined as in MultiCure}
}
\value{
a list containing 
\itemize{
\item [[1]]: deltaRImp, A single imputation of delta_R
\item [[2]]: YRImp, A single imputation of Y_R
}
}
\description{
The function UNEQUALCENSIMPUTECOXNESTEDWEIB will perform an imputation algorithm to handle unequal follow-up for recurrence and death. This function can be applied when we assume COX baseline hazards. Even though the main multistate cure model assumes Cox baseline hazards, this function (temporarily) uses Weibull baseline hazard assumptions to perform imputation of Y_R and delta_R. This function makes use of a rejection sampling algorithm.
}
