% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutcomeImputationWEIBINVERSION.R
\name{UNEQUALCENSIMPUTEWEIBINVERSION}
\alias{UNEQUALCENSIMPUTEWEIBINVERSION}
\title{UNEQUALCENSIMPUTEWEIBINVERSION}
\usage{
UNEQUALCENSIMPUTEWEIBINVERSION(datWIDE, beta, alpha, scale, shape,
  ImputeDat, TransCov)
}
\arguments{
\item{datWIDE}{defined as in MultiCure}

\item{beta}{A vector containing the most recent estimates of beta}

\item{alpha}{A vector containing the most recent estimates of alpha}

\item{scale}{A vector containing the most recent estimates of scale}

\item{shape}{A vector containing the most recent estimates of shape}

\item{ImputeDat}{This is a list with the following elements:
\itemize{
\item UnequalCens: A vector taking value 1 if the subject has unequal follow-up. Note: If subject is assumed cured in datWIDE, they are listed as UnequalCens = 0.
\item CovMissing: A matrix indicating which elements of Cov are missing. Not needed for this imputation.
\item CovImp: A list containing a single imputation of Cov
\item GImp: A vector with a recent single imputation of G
\item YRImp: A vector with a recent single imputation of Y_R
\item deltaRImp: A vector with a recent single imputation of delta_R
\item y: The integral of the target kernel over Yr0 to Yd
}}

\item{TransCov}{defined as in MultiCure}
}
\value{
a list containing 
\itemize{
\item [[1]]: deltaRImp, A single imputation of delta_R
\item [[2]]: YRImp, A single imputation of Y_R
}
}
\description{
The function UNEQUALCENSIMPUTEWEIB_TR will perform an imputation algorithm to handle unequal follow-up for recurrence and death. This function can be applied when we assume WEIBULL baseline hazards and 'T_R' is included in the model for Recurrence -> Death. Rather than imputing new values of YR using rejection sampling, this function performs imputation through inverting the survival function of the target distribution.
}
