% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarianceEM.R
\name{VarianceEM}
\alias{VarianceEM}
\title{VarianceEM}
\usage{
VarianceEM(fit, iternum, bootnum, datWIDE, Cov, ASSUME, TransCov, BASELINE,
  PENALTY = "None")
}
\arguments{
\item{fit}{Multistate cure model fit from MultiCure}

\item{iternum}{Number of iterations of the EM algorithm to apply to each bootstrap sample}

\item{bootnum}{Number of bootstrap samples}

\item{datWIDE}{A data frame with the following columns: 
\itemize{
\item Y_R, the recurrence event/censoring time
\item delta_R, the recurrence event/censoring indicator
\item Y_D, the death event/censoring time 
\item delta_D, the death event/censoring indicator
\item G, the cure status variable. This takes value 1 for known non-cured, 0 for "known" cured and NA for unknown cure status
}}

\item{Cov}{A matrix containing the covariates}

\item{ASSUME}{This variables indicates what equality assumptions we are making regarding the 24 and 14 transitions. The possible options are:
\itemize{
\item 'SameHazard': Lambda_14(t) = Lambda_24(t)
\item 'AllSeparate': No restrictions on Lambda_14(t) and Lambda_24(t)
\item  'ProportionalHazard': Lambda_14(t) = Lambda_24(t) exp(Beta0)
\item  'SameBaseHaz': Lambda^0_14(t) = Lambda^0_24(t), No restrictions on beta_14 and beta_24
}}

\item{TransCov}{a list with elements: Trans13, Trans24, Trans14, Trans34, PNonCure. Each list element is a vector containing the names of the variables in Cov to be used in the model for the corresponding transition. 13 is NonCured -> Recurrence, 24 is Cured -> Death, 14 is NonCured -> Death, 34 is Recurrence -> Death. PNonCure contains the names of the covariates for the logistic regression for P(NonCure).}

\item{BASELINE}{This variable indicates the assumptions about the baseline hazard form. This can take values 'weib' and 'cox'}

\item{PENALTY}{This variable indicates whether we are using any variable selection in the model fitting. Right now, the options are 'None' (no variable selection), 'Ridge' (ridge regression for all covariates in all models) and 'Lasso' (lasso for all covariates in all models, only implemented for Cox baseline hazards)}
}
\value{
OUT a matrix containing the following:
\itemize{
\item Estimate an estimate of the multistate cure model parameter
\item Variance an estimate of variance of the the multistate cure model parameter
}
}
\description{
The function VarianceEM performs variance estimation for the Multistate Cure Model when the model is fit using the EM Algorithm. Variances are estimated by fitting the multistate cure model (via EM) to bootstrap samples of the data.
}
\examples{
attach(SimulateMultiCure(type = "NoMissingness"))
Cov = data.frame(X1,X2)
VARS = names(Cov)
TransCov = list(Trans13 = VARS, Trans24 = VARS, Trans14 = VARS, Trans34 = VARS, PNonCure = VARS)
datWIDE = data.frame( Y_R, Y_D, delta_R , delta_D, G)
fit = MultiCure(iternum=100, datWIDE, Cov, ASSUME = "SameHazard", TransCov = TransCov, BASELINE = "weib") 
OUT = VarianceEM(fit,iternum=20, bootnum=50, datWIDE, Cov, ASSUME = "SameHazard", TransCov, BASELINE = "weib")
}
