% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarianceMCEM_NOBOOT.R
\name{VarianceMCEM_NOBOOT}
\alias{VarianceMCEM_NOBOOT}
\title{VarianceMCEM_NOBOOT}
\usage{
VarianceMCEM_NOBOOT(fit, datWIDE, CovImp, GImp, YRImp, deltaRImp, ASSUME,
  TransCov, BASELINE, PENALTY = "None")
}
\arguments{
\item{fit}{Multistate cure model fit from MultiCure}

\item{datWIDE}{A data frame with the following columns: 
\itemize{
\item Y_R, the recurrence event/censoring time
\item delta_R, the recurrence event/censoring indicator
\item Y_D, the death event/censoring time 
\item delta_D, the death event/censoring indicator
\item G, the cure status variable. This takes value 1 for known non-cured, 0 for "known" cured and NA for unknown cure status
}}

\item{CovImp}{A list with IMPNUM elements containing the imputations of Cov output from the function ProperDraws_MC}

\item{GImp}{A matrix with IMPNUM elements containing the imputations of G output from the function ProperDraws_MC}

\item{YRImp}{A matrix with IMPNUM elements containing the imputations of Y_R output from the function ProperDraws_MC}

\item{deltaRImp}{A matrix with IMPNUM elements containing the imputations of delta_R output from the function ProperDraws_MC}

\item{ASSUME}{This variables indicates what equality assumptions we are making regarding the 24 and 14 transitions. The possible options are:
\itemize{
\item 'SameHazard': Lambda_14(t) = Lambda_24(t)
\item 'AllSeparate': No restrictions on Lambda_14(t) and Lambda_24(t)
\item  'ProportionalHazard': Lambda_14(t) = Lambda_24(t) exp(Beta0)
\item  'SameBaseHaz': Lambda^0_14(t) = Lambda^0_24(t), No restrictions on beta_14 and beta_24
}}

\item{TransCov}{a list with elements: Trans13, Trans24, Trans14, Trans34, PNonCure. Each list element is a vector containing the names of the variables in Cov to be used in the model for the corresponding transition. 13 is NonCured -> Recurrence, 24 is Cured -> Death, 14 is NonCured -> Death, 34 is Recurrence -> Death. PNonCure contains the names of the covariates for the logistic regression for P(NonCure).}

\item{BASELINE}{This variable indicates the assumptions about the baseline hazard form. This can take values 'weib' and 'cox'}

\item{PENALTY}{This variable indicates whether we are using any variable selection in the model fitting. Right now, the options are 'None' (no variable selection), 'Ridge' (ridge regression for all covariates in all models) and 'Lasso' (lasso for all covariates in all models, only implemented for Cox baseline hazards)}

\item{bootnum}{Number of bootstrap samples used for each imputed dataset}
}
\value{
OUT a matrix containing the following:
\itemize{
\item Estimate an estimate of the multistate cure model parameter from Rubin's Rules
\item Variance an estimate of variance of the the multistate cure model parameter from Rubin's Rules
\item v The estimated degrees of freedom of the t-distribution of the parameter estimate from Rubin's Rules
}
}
\description{
The function VarianceMCEM_NOBOOT performs variance estimation for the Multistate Cure Model when the model is fit using a Monte Carlo EM Algorithm. For each imputed dataset, this function estimates parameter variances by fitting the multistate cure model the imputed dataset. Corresponding standard errors come are extracted from the corresponding logistic and proportional hazards model fits. Then, Rubin's rules are used to obtain a single set of parameter estimates and standard errors across imputed datasets. Note: This function has not yet been implemented for PENALTY values other than 'None'. Important!!!: The function ProperDraws_MC should be used to obtain the imputed datasets to be used in this function. This ensures that the imputed datasets are (roughly) proper imputations and that Rubin's rules can then be applied.
}
\details{
This function provides parameter estimates and estimated variances. The parameter estimates are obtained using Rubin's rules, but an alternative estimate of the multistate cure model parameter can be obtained by averaging the parameter estimates from the last few iterations of the model fitting algorithm. In our experience, we found that the approach that averages across the last few iterations (rather than estimated using Rubin's rules) provides a better estimate of the parameter of interest.
}
\author{
Lauren J Beesley, \email{lbeesley@umich.edu}
}
