% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VISUALIZEDATA.R
\name{VISUALIZEDATA}
\alias{VISUALIZEDATA}
\title{VISUALIZEDATA}
\usage{
VISUALIZEDATA(datWIDE)
}
\arguments{
\item{datWIDE}{A data frame with the following columns: 
\itemize{
\item Y_R, the recurrence event/censoring time
\item delta_R, the recurrence event/censoring indicator
\item Y_D, the death event/censoring time 
\item delta_D, the death event/censoring indicator
}}
}
\description{
This function uses RShiny to plot Kaplan-Meier plots for time to recurrence and time to death. This function can also be used to visualize the occurrence of unequal follow-up and when the recurrence event/censorings and death event/censorings occur.
}
\examples{
attach(SimulateMultiCure(type = "NoMissingness"))
datWIDE = data.frame( Y_R, Y_D, delta_R , delta_D, G)
VISUALIZEDATA(datWIDE)
### With Unequal Censoring
attach(SimulateMultiCure(type = "UnequalCensoring"))
datWIDE = data.frame( Y_R, Y_D, delta_R , delta_D, G)
VISUALIZEDATA(datWIDE)
}
