% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATEOCCUPANCYWEIB.R
\name{STATEOCCUPANCYWEIB}
\alias{STATEOCCUPANCYWEIB}
\title{STATEOCCUPANCYWEIB}
\usage{
STATEOCCUPANCYWEIB(times, TransCov, Cov, beta, alpha, scale, shape)
}
\arguments{
\item{times}{number of iterations for the EM or MCEM algorithm}

\item{TransCov}{a list with elements: Trans13, Trans24, Trans14, Trans34, PNonCure. Each list element is a vector containing the names of the variables in Cov to be used in the model for the corresponding transition. 13 is NonCured -> Recurrence, 24 is Cured -> Death, 14 is NonCured -> Death, 34 is Recurrence -> Death. PNonCure contains the names of the covariates for the logistic regression for P(NonCure).}

\item{Cov}{A dataframe with columns corresponding to covariates used in the model fit. The rows should correspond to the covariate values for which we will make our predictions.}

\item{beta}{Estimate from multistate cure model fit}

\item{alpha}{Estimate from multistate cure model fit}

\item{scale}{Estimate from multistate cure model fit}

\item{shape}{Estimate from multistate cure model fit}
}
\description{
This function uses RShiny to plot predicted state occupancy probabilities, overall survival probabilities, event-free probabilities, and the transition densities for a particular set of covariate values. These probabilities are estimated based on a Multistate cure model fit with WEIBULL baseline hazards. The estimated probabilities can incorporate only baseline information OR can incorporate baseline information and some followup information after baseline. This function cannot be applied for prediction when recurrence time is included in the model from recurrence to death.
}
\author{
Lauren J Beesley, \email{lbeesley@umich.edu}
}
